+++ packages/common/src/util/index.js
@@ -3,5 +3,6 @@ export * from './references.js';
 
 export { default as parseDate } from './parse-date.js';
 export { default as throwError } from './throw-error.js';
+export { sanitizeError } from './sanitize-error.js';
 export { encode, decode } from './base64.js';
 export { uuid } from './uuid.js';