+++ packages/googledrive/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-googledrive
 
+## 3.1.3 - 17 July 2026
+
+### Patch Changes
+
+- [#1706](https://github.com/OpenFn/adaptors/issues/1706): Sanitize
+  credential-bearing client errors before propagation.
+- Updated dependencies
+  - @openfn/language-common@3.4.0
+
 ## 3.1.2 - 30 June 2026
 
 ### Patch Changes